+++ src/Python/Visualization/open3d_visualization.h
@@ -26,7 +26,7 @@
 
 #pragma once
 
-#include <Python/open3d_python.h>
+#include <Python/py_open3d.h>
 
 void pybind_renderoption(py::module &m);
 void pybind_viewcontrol(py::module &m);
