+++ src/Python/Visualization/open3d_visualization_trampoline.h
@@ -26,7 +26,7 @@
 
 #pragma once
 
-#include <Python/open3d_python.h>
+#include <Python/py_open3d.h>
 #include <Open3D/Visualization/Visualizer/Visualizer.h>
 #include <Open3D/Visualization/Visualizer/ViewControl.h>
